:zap: Make release ready
Arguments have been updated for full command line access and output
saving commands have been adapted to give full all useful information.
A new output file `energies.txt` will be created when `-a` is used, this
contains the energies of each alanine scan. The format of all output
files is now tab-separated, with blank places indicated with '--' to
ensure compatibility with most text parsing tools. The README has been
updated to include usage guidance and some references for the algorithm.
A full description of the algorithm is nto included at this stage.